denidin-app/requirements.txt
# DeniDin WhatsApp AI Chatbot - Python Dependencies
# Phase 1: Infrastructure Setup

# Green API (WhatsApp Business API integration)
whatsapp-chatbot-python>=0.5.1
whatsapp-api-client-python>=0.0.53
whatsapp-chatgpt-python>=0.0.1

# OpenAI API
openai>=1.12.0
tiktoken>=0.5.0  # Token counting for conversation limits

# Memory & Vector Storage (Feature 002+007)
chromadb>=0.4.22

# Media Processing (Feature 003)
PyMuPDF>=1.23.0           # PDF to image conversion (fitz)
python-docx>=1.0.0        # DOCX text extraction
Pillow>=10.0.0            # Image processing utilities

# Configuration & Data
PyYAML>=6.0

# Retry logic for API calls
tenacity>=8.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.12.0

# Code Quality
mypy>=1.0.0
pylint>=3.0.0
